move agents around + fix episodes
