+++ projects/stretch_ovmm/eval_episode.py
@@ -4,7 +4,7 @@ from typing import Optional, Tuple
 import click
 import rospy
 
-from home_robot.agent.hierarchical.pick_and_place_agent import PickAndPlaceAgent
+from home_robot.agent.ovmm_agent.pick_and_place_agent import PickAndPlaceAgent
 from home_robot.motion.stretch import STRETCH_HOME_Q
 from home_robot_hw.env.stretch_pick_and_place_env import (
     REAL_WORLD_CATEGORIES,
