+++ src/home_robot/home_robot/agent/ovmm_agent/__init__.py
@@ -1,1 +1,2 @@
 from .ovmm_agent import OpenVocabManipAgent, Skill, get_skill_as_one_hot_dict
+from .pick_and_place_agent import PickAndPlaceAgent
